chart: amber auto-sell segments on the SOC line; merge live-data + diagnostics into /system
The sim always credited battery-full PV overflow at spot (autoExportWh)
— now it also reports the per-slot wattage, and the chart paints those
stretches of the SOC line in solar amber with the power in the tooltip,
so "flat at 100%" visibly means "selling the surplus". Legend wording
becomes theme-neutral ("stronger colour = more power" — higher opacity
is darker on light, brighter on dark).

Live data and Diagnostics merge into one /system page (raw sensor truth
on top, controller internals below); old URLs redirect, nav shrinks to
five items.
