+++ frontend/src/components/system/DiagnosticsCards.scss
