+++ frontend/src/components/system/LiveReadings.scss
