+++ frontend/src/routes/system.scss
@@ -1,0 +1,8 @@
+.system__heading {
+  font-size: 11.5px;
+  font-weight: 650;
+  letter-spacing: 0.1em;
+  text-transform: uppercase;
+  color: var(--ink-2);
+  margin: 1.6rem 0 0.7rem;
+}
